Add support for Digital Ocean by enhancing exec auth plugin
