+++ lib/auth/gcp.js
@@ -22,28 +22,28 @@ class GoogleCloudPlatform {
   }
 
   refresh_token() {
-    let promise, cancel;
+    let promise, cancel = function (){};
     if (os.platform() === 'browser') {
       console.error('Refreshing GCP token in the browser is not supported! Please refresh token manually via \'gcloud container clusters get-credentials\'. ');
       promise = Promise.resolve(this.auth_provider.access_token);
-    }
-    promise = new Promise((resolve, reject) => {
-      const process = execFile(this.auth_provider.cmd_path, this.auth_provider.cmd_args , (error, stdout, stderr) => {
-        cancel = function() {};
-        if (error) {
-          reject(error);
-        }
-        resolve(stdout);
+    } else {
+      promise = new Promise((resolve, reject) => {
+        const process = execFile(this.auth_provider.cmd_path, this.auth_provider.cmd_args , (error, stdout, stderr) => {
+          cancel = function() {};
+          if (error) {
+            reject(error);
+          }
+          resolve(stdout);
+        });
+        cancel = function() { process.kill() };
       });
-      cancel = function() { process.kill() };
-    });
-    promise = promise.then( response => {
-      const json = JSON.parse(response);
-      this.auth_provider.expiry = Date.parse(JSONPath(this.auth_provider.expiry_key, json)[0]);
-      this.auth_provider.access_token = JSONPath(this.auth_provider.token_key, json)[0];
-      return this.auth_provider.access_token;
-    });
-
+      promise = promise.then( response => {
+        const json = JSON.parse(response);
+        this.auth_provider.expiry = Date.parse(JSONPath(this.auth_provider.expiry_key, json)[0]);
+        this.auth_provider.access_token = JSONPath(this.auth_provider.token_key, json)[0];
+        return this.auth_provider.access_token;
+      });
+    }
     return { promise, cancellation : () => cancel() };
   }
