+++ lib/client.js
@@ -2,7 +2,7 @@
 
 const crypto              = require('crypto'),
       get                 = require('./http-then').get,
-      ExecAuth            = require('./auth/exec-auth'),
+      ExecAuth            = require('./auth/exec'),
       OpenIdConnect       = require('./auth/oidc'),
       GoogleCloudPlatform = require('./auth/gcp'),
       URI                 = require('urijs');
