+++ lib/config/user.js
@@ -26,7 +26,7 @@ class User {
 
   constructor({ name, token, username, password, 'client-certificate': certificatePath,
       'client-certificate-data': certificateBase64, 'client-key': keyPath,
-      'client-key-data': keyBase64, 'auth-provider': auth_provider, 'exec': exec }) {
+      'client-key-data': keyBase64, 'auth-provider': auth_provider, exec }) {
     if (typeof name === 'undefined') {
       throw Error('User name must be defined!');
     }
@@ -34,7 +34,7 @@ class User {
       this.auth_provider = new AuthProvider(Object.assign({ name: auth_provider.name, config: auth_provider.config }));
     }
     if (typeof exec !== 'undefined') {
-      //for now let's treat exec like an Auth Provider
+      // for now let's treat exec like an AuthProvider
       this.auth_provider = new AuthProvider(Object.assign({ name: 'exec', config: exec }));
     }
     this.name = name;
@@ -135,13 +135,12 @@ class GoogleCloudPlatformAuthProvider {
 }
 
 class ExecAuthProvider {
-  constructor({ 'apiVersion': api_version, 'args': args, 'command': command,
-      'env': env}) {
+  constructor({ 'apiVersion': api_version, args, command, env }) {
     this.api_version = api_version;
     this.args = args;
     this.command = command;
     this.env = env;
-    this.expiry = Date.now()
+    this.expiry = Date.now();
   }
 
   get auth_token() {
